1. 一对一及联新增、修改、查询、删除功能 2. 增加ID自动生成类型snowflake

+++ src/main/java/com/rj/btp/biz/hrms/employee/reopsitory/EmployeeRepository.java
@@ -5,5 +5,5 @@ import com.rj.btp.framework.mapper.repository.BaseRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeRepository extends BaseRepository<Employee, Integer> {
+public interface EmployeeRepository extends BaseRepository<Employee, Long> {
 }
